refactor(chat): a conversation is a component, not a page
The reader is meant to dock this conversation rather than navigate away from
the document to reach it, and a page-level default export cannot be mounted
anywhere. `ChatConversation` now holds the thread, the composer and the session
list; `variant="docked"` drops the session list, the back button and the `?q=`
prefill, which are the only things that were ever about being a page. Chat.tsx
is the route around it.

The roadmap's premise for this rung is corrected in the same change: the global
slide-over it names is unreachable -- `setChatPanelOpen(true)` exists nowhere,
so App.tsx mounts the whole Chat page into 450px behind a flag no code sets.

What a docked conversation is scoped to is still open and is recorded as such:
`chatScope`, `chatSelectedDocId`, `chatMessages` and `activeChatSessionId` are
single global values, so docking as-is would re-scope the Ask page.

Verified in the browser: an answer streams with its receipt and chips, and
citations still land (5/5 in view, twice).
